+++ src/lib.rs
@@ -1,5 +1,5 @@
 //! # Particular
-//! 
+//!
 //! Particular is a crate providing a simple way to simulate N-body gravitational interaction of particles in Rust.
 //!
 //! ## Goals
@@ -15,16 +15,30 @@
 //!
 //! The API to setup a simulation is straightforward:
 //!
-//! ## Implementing the [`Particle`] trait
+//! ## Implementing the [`Particle`](particle::Particle) trait
 //!
-//! #### Deriving:
+//! #### Attribute macro or deriving:
 //!
-//! Used in most cases, when the type has fields named `position` and `mu`.
+//! Used in most cases, when the type has fields named `position` and `mu`:
 //! ```
-//! # use particular::prelude::Particle;
+//! # use particular::prelude::*;
+//! # use glam::Vec3;
+//! #
+//! #[particle(3)]
+//! pub struct Body {
+//!     position: Vec3,
+//!     mu: f32,
+//! //  ...
+//! }
+//! ```
+//! This is equivalent to:
+//!
+//! ```
+//! # use particular::prelude::*;
 //! # use glam::Vec3;
 //! #
 //! #[derive(Particle)]
+//! #[dim(3)]
 //! pub struct Body {
 //!     position: Vec3,
 //!     mu: f32,
@@ -37,7 +51,7 @@
 //! ```
 //! # const G: f32 = 1.0;
 //! #
-//! # use particular::Particle;
+//! # use particular::prelude::*;
 //! # use glam::Vec3;
 //! #
 //! struct Body {
@@ -47,7 +61,7 @@
 //! }
 //!
 //! impl Particle for Body {
-//!     type Vector = Vec3;
+//!     type Vector = VectorDescriptor<3, Vec3>;
 //!
 //!     fn position(&self) -> Vec3 {
 //!         self.position
@@ -59,16 +73,15 @@
 //! }
 //! ```
 //! ## Setting up the simulation
-//! Using the type implementing [`Particle`], create a [`ParticleSet`] that will contain the particles.
+//! Using the type implementing [`Particle`](particle::Particle), create a [`ParticleSet`](particle_set::ParticleSet) that will contain the particles.
 //!
 //! Particles are stored in two vectors, `massive` or `massless`, depending on if they have mass or not.
 //! This allows optimizations in the case of massless particles (which represents objects that do not need to affect other objects, like a spaceship).
 //! ```
-//! # use particular::prelude::Particle;
-//! # use particular::ParticleSet;
+//! # use particular::prelude::*;
 //! # use glam::Vec3;
 //! #
-//! # #[derive(Particle)]
+//! #[particle(3)]
 //! # pub struct Body {
 //! #     position: Vec3,
 //! #     mu: f32,
@@ -77,43 +90,46 @@
 //! # let position = Vec3::ONE;
 //! # let mu = 1E5;
 //! #
-//! let mut particle_set = ParticleSet::new();
 //! // If the type cannot be inferred, use the turbofish syntax:
-//! // let mut particle_set = ParticleSet::<Body>::new();
+//! let mut particle_set = ParticleSet::<Body>::new();
+//! // Otherwise:
+//! let mut particle_set = ParticleSet::new();
 //!
 //! particle_set.add(Body { position, mu });
 //! ```
 //! ## Computing and using the gravitational acceleration
-//! Finally, use the [`result`](ParticleSet::result) method of [`ParticleSet`], which returns an iterator over a mutable reference to the `Particle` and its computed gravitational acceleration.
+//! Finally, use the [`result`](particle_set::ParticleSet::result) method of [`ParticleSet`](particle_set::ParticleSet), which returns an iterator over a mutable reference to the `Particle` and its computed gravitational acceleration.
 //! ```
-//! # use particular::prelude::Particle;
-//! # use particular::ParticleSet;
+//! # use particular::prelude::*;
 //! # use glam::Vec3;
 //! #
 //! # const DT: f32 = 1.0 / 60.0;
 //! #
-//! # #[derive(Particle)]
+//! #[particle(3)]
 //! # pub struct Body {
 //! #     position: Vec3,
 //! #     velocity: Vec3,
 //! #     mu: f32,
 //! # }
 //! # let mut particle_set = ParticleSet::<Body>::new();
-//! for (particle, acceleration) in particle_set.result() {
+//! for (acceleration, particle) in particle_set.result() {
 //!     particle.velocity += acceleration * DT;
 //!     particle.position += particle.velocity * DT;
 //! }
 //! ```
 
-mod particle;
-mod particle_set;
-mod vector;
+pub mod particle;
+pub mod particle_set;
+pub mod vector;
 
-pub use particle::*;
-pub use particle_set::*;
-pub use vector::*;
+pub mod particular_derive {
+    pub use crate::vector::VectorDescriptor;
+    pub use particular_derive::{particle, Particle};
+}
 
 pub mod prelude {
-    pub use crate::{vector_normed, NormedVector, Particle, ParticleSet};
-    pub use particular_derive::Particle;
+    pub use crate::particle::Particle;
+    pub use crate::particle_set::ParticleSet;
+    pub use crate::particular_derive::*;
+    pub use crate::vector::{Descriptor, VectorDescriptor};
 }